refactor the store structure
