+++ src/backend/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/atom/MarketAtomArchiveService.kt
@@ -26,4 +26,30 @@
 
 package com.tencent.devops.store.service.atom
 
-interface EeMarketAtomService
+import com.tencent.devops.common.api.pojo.Result
+import com.tencent.devops.store.pojo.atom.AtomEnvRequest
+import com.tencent.devops.store.pojo.atom.GetAtomConfigResult
+import com.tencent.devops.store.pojo.atom.enums.ReleaseTypeEnum
+
+interface MarketAtomArchiveService {
+
+    fun getTaskJsonStr(projectCode: String, atomCode: String, version: String): String
+
+    fun verifyAtomPackageByUserId(
+        userId: String,
+        projectCode: String,
+        atomCode: String,
+        version: String,
+        releaseType: ReleaseTypeEnum?,
+        os: String?
+    ): Result<Boolean>
+
+    fun verifyAtomTaskJson(
+        userId: String,
+        projectCode: String,
+        atomCode: String,
+        version: String
+    ): Result<GetAtomConfigResult?>
+
+    fun updateAtomEnv(userId: String, atomId: String, atomEnvRequest: AtomEnvRequest): Result<Boolean>
+}
